Modify j,k behavior with wrapped lines
When using j or k alone to navigate across wrapped lines, move based on
the lines displayed (i.e. take the wrapping into account)

When using j or k with a count, move based on the lines as defined in
the file (i.e. do NOT take the wrapping into account)

Fixes #6