Add new Vim command :Ag
The new :Ag command works like :grep (but using ag instead of grep).

Note: If ag is not present on the system, this command will fail.

This new :Ag command works by temporarily changing the values of:

- grepprg
- grepformat

to use ag, performing the search, and then restoring these variables to
their original values.

See #95