Add support for project specific .vdebug_path_maps
The .vdebug_path_maps file contains the local path map for vdebug on the
first line (and optionally the remote path map on the second line).

Vim will check for .vdebug_path_maps in the current directory when
launched. If the file is not present, the default path maps set in
init.vim will be used.

See #167, #166